+++ mysql_statsd/thread_base.py
@@ -1,0 +1,14 @@
+import threading
+
+
+class ThreadBase(threading.Thread):
+    run = True
+
+    def __init__(self, queue, **kwargs):
+        threading.Thread.__init__(self)
+        self.queue = queue
+        if getattr(self, 'configure', None):
+            self.configure(kwargs)
+
+    def stop(self):
+        self.run = False
